requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain
langchain-community
faiss-cpu
sentence-transformers
python-docx
langchain-google-genai
langchain-huggingface
